changes enable setup.py
